Move sirene.py to provider module.
