+++ erp/provider/entreprise.py
@@ -1,0 +1,1 @@
+# XXX https://entreprise.data.gouv.fr/api/sirene/v1/full_text/akei%20jacou?per_page=5&page=1
